Lille ændring til en returntype på en metode

+++ GUIbaseret/src/sample/domain/RoadBuilder.java
@@ -29,7 +29,7 @@ public class RoadBuilder {
         }
     }
 
-    public static double getInventoryCount() {
+    public static int getInventoryCount() {
         return inventoryCount;
     }
